Reduce diff and cleanup implementation
