+++ vertx-web-auth-common/src/main/java/io/vertx/ext/auth/common/AbstractUserContext.java
@@ -12,10 +12,16 @@ import io.vertx.ext.web.common.HttpException;
 
 public abstract class AbstractUserContext implements UserContextInternal {
 
+  public static final String USER_SWITCH_KEY = "__vertx.user-switch-ref";
+
   private static final Logger LOG = LoggerFactory.getLogger(AbstractUserContext.class);
 
   protected AuthenticationContext ctx;
   protected User user;
+
+  public AbstractUserContext(AuthenticationContext ctx) {
+    this.ctx = ctx;
+  }
 
   @Override
   public void setUser(User user) {
@@ -25,6 +31,30 @@ public abstract class AbstractUserContext implements UserContextInternal {
   @Override
   public User get() {
     return user;
+  }
+
+  @Override
+  public UserContext loginHint(String loginHint) {
+    final Session session = ctx.session();
+
+    if (session == null) {
+      if (loginHint == null) {
+        // Fine, we don't need a session
+        return this;
+      }
+      // we always need a session, otherwise we can't track the state of the previous user
+      throw new IllegalStateException("SessionHandler not seen in the route. Sessions are required to keep the state");
+    }
+
+    if (loginHint == null) {
+      // we're removing the hint if present
+      session.remove("login_hint");
+    } else {
+      session
+        .put("login_hint", loginHint);
+    }
+
+    return this;
   }
 
   @Override
@@ -39,6 +69,21 @@ public abstract class AbstractUserContext implements UserContextInternal {
   @Override
   public Future<Void> refresh(String redirectUri) {
     Objects.requireNonNull(redirectUri, "redirectUri cannot be null");
+
+    if (user == null) {
+      // we need to ensure that we already had a user, otherwise we can't switch
+      LOG.debug("Impersonation can only occur after a complete authn flow.");
+      return Future.failedFuture(new HttpException(401));
+    }
+
+    final Session session = ctx.session();
+
+    if (session != null) {
+      // From now on, we're changing the state
+      session
+        // force a session id regeneration to protect against replay attacks
+        .regenerateId();
+    }
 
     // remove user from the context
     this.user = null;
@@ -56,11 +101,6 @@ public abstract class AbstractUserContext implements UserContextInternal {
   }
 
   @Override
-  public Future<Void> logout() {
-    return logout("/");
-  }
-
-  @Override
   public Future<Void> impersonate() {
     if (!ctx.request().method().equals(HttpMethod.GET)) {
       // we can't automate a redirect to a non-GET request
@@ -72,6 +112,33 @@ public abstract class AbstractUserContext implements UserContextInternal {
   @Override
   public Future<Void> impersonate(String redirectUri) {
     Objects.requireNonNull(redirectUri, "redirectUri cannot be null");
+
+    if (user == null) {
+      // we need to ensure that we already had a user, otherwise we can't switch
+      LOG.debug("Impersonation can only occur after a complete authn flow.");
+      return Future.failedFuture(new HttpException(401));
+    }
+
+    final Session session = ctx.session();
+
+    if (session == null) {
+      // we always need a session, otherwise we can't track the state of the previous user
+      LOG.debug("SessionHandler not seen in the route. Sessions are required to keep the state");
+      return Future.failedFuture(new HttpException(500));
+    }
+
+    if (session.get(USER_SWITCH_KEY) != null) {
+      // we always need a session, otherwise we can't track the state of the previous user
+      LOG.debug("Impersonation already in place");
+      return Future.failedFuture(new HttpException(400));
+    }
+
+    // From now on, we're changing the state
+    session
+      // move the user out of the context (yet keep it in the session, so we can roll back
+      .put(USER_SWITCH_KEY, user)
+      // force a session id regeneration to protect against replay attacks
+      .regenerateId();
 
     // remove the current user from the context to avoid any further access
     this.user = null;
@@ -101,6 +168,43 @@ public abstract class AbstractUserContext implements UserContextInternal {
   public Future<Void> restore(String redirectUri) {
     Objects.requireNonNull(redirectUri, "redirectUri cannot be null");
 
+    if (user == null) {
+      // we need to ensure that we already had a user, otherwise we can't switch
+      LOG.debug("Impersonation can only occur after a complete authn flow.");
+      return Future.failedFuture(new HttpException(401));
+    }
+
+    final Session session = ctx.session();
+
+    if (session == null) {
+      // we always need a session, otherwise we can't track the state of the previous user
+      LOG.debug("SessionHandler not seen in the route. Sessions are required to keep the state");
+      return Future.failedFuture(new HttpException(500));
+    }
+
+    if (session.get(USER_SWITCH_KEY) == null) {
+      // we always need a session, otherwise we can't track the state of the previous user
+      LOG.debug("No previous impersonation in place");
+      return Future.failedFuture(new HttpException(400));
+    }
+
+    // From now on, we're changing the state
+    User previousUser = session.get(USER_SWITCH_KEY);
+
+    session
+      // move the user out of the context (yet keep it in the session, so we can rollback
+      .remove(USER_SWITCH_KEY);
+    // remove the previous hint
+    session
+      .remove("login_hint");
+
+    session
+      // force a session id regeneration to protect against replay attacks
+      .regenerateId();
+
+    // restore it to the context
+    this.user = previousUser;
+
     // we should redirect the UA so this link becomes invalid
     return ctx.response()
       // disable all caching
@@ -114,8 +218,19 @@ public abstract class AbstractUserContext implements UserContextInternal {
   }
 
   @Override
+  public Future<Void> logout() {
+    return logout("/");
+  }
+
+  @Override
   public Future<Void> logout(String redirectUri) {
     Objects.requireNonNull(redirectUri, "redirectUri cannot be null");
+
+    final Session session = ctx.session();
+    // clear the session
+    if (session != null) {
+      session.destroy();
+    }
 
     // clear the user
     user = null;
@@ -134,7 +249,13 @@ public abstract class AbstractUserContext implements UserContextInternal {
 
   @Override
   public void clear() {
+    final Session session = ctx.session();
+    // clear the session
+    if (session != null) {
+      session.destroy();
+    }
+
+    // clear the user
     user = null;
   }
-
 }
