+++ vertx-web/src/main/java/io/vertx/ext/web/handler/impl/WebHTTPAuthorizationHandler.java
@@ -1,7 +1,6 @@
 package io.vertx.ext.web.handler.impl;
 
 import io.vertx.core.http.HttpHeaders;
-import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.authentication.AuthenticationProvider;
 import io.vertx.ext.auth.common.handler.impl.HTTPAuthorizationHandler;
 import io.vertx.ext.web.RoutingContext;
@@ -11,10 +10,6 @@ public abstract class WebHTTPAuthorizationHandler<T extends AuthenticationProvid
 
   public WebHTTPAuthorizationHandler(T authProvider, Type type, String realm) {
     super(authProvider, type, realm);
-  }
-
-  public void postAuthentication(RoutingContext ctx, User authenticatedUser) {
-    ctx.next();
   }
 
   // TODO remove duplicated code from WebAuthenticationHandlerImpl
