+++ vertx-web/src/main/java/io/vertx/ext/web/impl/RoutingContextDecorator.java
@@ -1,7 +1,6 @@
 package io.vertx.ext.web.impl;
 
 import io.vertx.codegen.annotations.Nullable;
-
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.MultiMap;
