+++ vertx-web/src/test/java/io/vertx/ext/web/handler/ChainAuthMixHandlerTest.java
@@ -8,7 +8,6 @@ import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.jwt.JWTAuth;
 import io.vertx.ext.auth.jwt.JWTAuthOptions;
 import io.vertx.ext.web.WebTestBase;
-import io.vertx.ext.web.handler.HttpException;
 import io.vertx.ext.web.sstore.LocalSessionStore;
 import org.junit.Test;
